requirements.txt
Flask==0.10.1
werkzeug==0.14.1

# Application server for both development and production.
gunicorn==19.4.5

# Testing and static analysis.
pytest==5.1.0
pytest-cov==2.7.1
flake8==3.7.8
mock==1.3.0

# CLI.
Click==6.4

# Data and workers.
psycopg2==2.8.3
Flask-SQLAlchemy==2.1
SQLAlchemy==1.2.0
redis==3.3.7
celery==4.3.0

# Forms.
Flask-WTF==0.9.5
WTForms-Components==0.9.7

# Payments.
stripe==1.32.0

# Utils.
fake-factory==0.5.7

# Extensions.
flask-debugtoolbar==0.10.0
Flask-Mail==0.9.1
Flask-Login==0.3.2
Flask-Limiter==0.9.3